Move plot_slices from utility module into zdcsummary, so no pkg install is required
